Walls method
Currently doesn't work, due to the available function not being implemented correctly, maybe move out of the class and call it independently. Visualise function also broken too, as the lines don't line up. Need to have a travelled, as otherwise it will make loops in the maze using this algorithm. Need to rework the choice at which direction to go, as currently it returns an Integer rather than a String to iterate through.