Refactor + Features + Optimisations
Optimised the depth first algorithm. Now greater than ~10x the speed it was before. It was mostly the drawing which was being done inefficiently. There are potential ways to increase the speed, for example joining connected lines together and just drawing it once. To do this perhaps wait until there is a break then render the new big line.

ie) lastx, lasty. if not line, draw lastx, lasty, currentx, currenty. else continue.

This would only work for horizontal lines though, and would have to add a seperate counter for the vertical ones.

Need to add entrances.

Added a solution visualiser (may not be the best path if random is used).

Added starting and ending point options so you can choose where to start and end.

Made the final screen resizeable such that the maze resizes with the screen.

Can show or hide the solution with Return.

When it reaches the end it goes back 3 nodes in the sequence to hopefully make the end more out of the way. ie only one path to it.

Added a chance for a travelled node to be ignored, creating mazes with multiple solutions.

Refactored the code to look and read nicer.

Removed duplicate files.